requirements.txt
flask
Flask-Cors
pymongo
python-dotenv
pydantic
cloudscraper
python-dotenv